forth stage cleanup
